Перенес интерфейс для самостоятельных объектов метаданных
