+++ src/main/java/com/github/_1c_syntax/bsl/mdo/Command.java
@@ -24,5 +24,5 @@ package com.github._1c_syntax.bsl.mdo;
 /**
  * Интерфейс объектов команд объектов и общих команд
  */
-public interface Command extends MDObject, ModuleOwner {
+public interface Command extends MD, ModuleOwner {
 }
