+++ src/main/java/com/github/_1c_syntax/bsl/mdo/Form.java
@@ -26,7 +26,9 @@ import com.github._1c_syntax.bsl.mdo.support.FormType;
 /**
  * Интерфейс форм объектов и общих форм
  */
-public interface Form extends MDObject, ModuleOwner {
+public interface Form extends MD
+// todo  , ModuleOwner
+{
   /**
    * Тип формы
    */
