+++ src/main/java/com/github/_1c_syntax/mdclasses/mdo/attributes/AbstractMDOAttribute.java
@@ -21,7 +21,8 @@
  */
 package com.github._1c_syntax.mdclasses.mdo.attributes;
 
-import com.github._1c_syntax.bsl.mdo.MDChildObject;
+import com.github._1c_syntax.bsl.mdo.Attribute;
+import com.github._1c_syntax.bsl.mdo.MDChild;
 import com.github._1c_syntax.bsl.mdo.support.AttributeKind;
 import com.github._1c_syntax.bsl.mdo.support.IndexingType;
 import com.github._1c_syntax.bsl.types.MdoReference;
@@ -41,7 +42,7 @@ import lombok.ToString;
 @EqualsAndHashCode(callSuper = true, onlyExplicitlyIncluded = true)
 @ToString(callSuper = true, onlyExplicitlyIncluded = true)
 @NoArgsConstructor
-public abstract class AbstractMDOAttribute extends AbstractMDObjectBase implements MDChildObject {
+public abstract class AbstractMDOAttribute extends AbstractMDObjectBase implements Attribute, MDChild {
 
   /**
    * Вид атрибута
